% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/bt88.03.704-package.R
\docType{package}
\name{bt88.03.704-package}
\alias{bt88.03.704-package}
\title{float's utility functions

Some useful little functions that make life easier.}
\description{
float's utility functions

Some useful little functions that make life easier.
}

