% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{count_significant}
\alias{count_significant}
\title{count significant genes}
\usage{
count_significant(model, coef, cutoff = 0.05, column = "adj.P.Val")
}
\arguments{
\item{model}{\code{MArrayLM} object fit by \code{limma}}

\item{coef}{\code{integer} index of contrast to sort genes by}

\item{cutoff}{significance level}

\item{column}{which column should the cutoff be applied to}
}
\value{
number of significant probesets
}
\description{
count the number of genes signifcant at a certain level
}
\author{
float
}

