% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expressionLines}
\alias{expressionLines}
\title{Expression line plots}
\usage{
expressionLines(data, normalized = NULL, n = 10, faulty = NULL,
  types = c("raw", "normalized"))
}
\arguments{
\item{data}{expression dataset, object for which \code{exprs} returns a matrix of expression values}

\item{normalized}{alternative expression dataset to compare with \code{data}}

\item{n}{number of rows to plot, or character vector with rownames to be plotted}

\item{faulty}{numeric index vector or column names of chips that should be highlighted in the plots}
}
\value{
ggplot2 plot object
}
\description{
Draw a series of line plots of expression values. \code{expressionLines} line plots of the first \code{n} probes from array data \code{data}. If a second matching dataset is given lines for the corresponding expressions in this dataset are overlayed. This maybe helpful to compare a set of expressions in the raw and normalized versions of a dataset.
}
\author{
float
}

