% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edger.R
\name{filterDGE}
\alias{filterDGE}
\title{Filter DGE}
\usage{
filterDGE(dge, mCt = 1.5, atleast = 2, topM = 8000)
}
\arguments{
\item{dge}{DGEList with transcript reads}

\item{mCt}{minimum count}

\item{atleast}{number of samples for which count has to exceed \code{mCT}}

\item{topM}{number transcripts with largest MAD to be selected for further processing}
}
\value{
DGEList with filtered transcripts
}
\description{
Apply non-specific filtering to DGEList
}
\author{
Florian Klinglmueller
}

