% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edger.R
\name{fit.edger}
\alias{fit.edger}
\title{edgeR wrapper}
\usage{
fit.edger(data, MM, CM, ...)
}
\arguments{
\item{data}{a DGEList object}

\item{MM}{a model matrix}

\item{CM}{a contrast matrix}

\item{...}{additional parameters passed to glmLRT}
}
\value{
a DGELRT object
}
\description{
Wrapper script that fits an edgeR model and does linear contrast tests.
}
\examples{
## setting up some random data
data <- matrix(rpois(600,200),ncol=6)
dge <- DGEList(counts=data)
dge <- estimateDisp(dge)

## a design matrix
design <- data.frame('Treatment' = factor(rep(0:1,each=3)))
mm <- model.matrix(~0+Treatment,design)

## a contrast matrix
colnames(mm) <- c('control','treatment')
cm <- makeContrasts('control-treatment',levels=mm)

## fit and test
out <- fit.edger(dge,mm,cm)

myTopTags(out,3)
}
\author{
Florian Klinglmueller
}

