% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatslide.R
\name{heatslide}
\alias{heatslide}
\title{Heatmap with dotplot like visualization of effect sizes}
\usage{
heatslide(mat, stat, pheno, genenames = NULL,
  hcols = colorRampPalette(c("blue", "white", "red"))(32),
  lcols = rainbow(length(levels(pheno))), scaleValues = c("row", "column",
  "none"), scaler = "linear", slidetitle = "Log (Base 2) Foldchange",
  rowSort = TRUE, colSort = TRUE)
}
\arguments{
\item{mat}{Matrix containing the values to be plotted in the heatmap}

\item{stat}{Vector of values that should be plotted by the dotplot beside
each line of the heatmap}

\item{pheno}{Vector giving the group labels for each column of the matrix}

\item{genenames}{Character vector giving genenames (or corresponding annotation)}

\item{hcols}{Colours to be used in the heatmap}

\item{lcols}{Colours to indicate the column labels}

\item{scaleValues}{Scale expression values by either 'row', 'column', or 'none'}

\item{scaler}{if "linear" use scale and remove mean and divide by sd, if "rank" transform to ranks}

\item{slidetitle}{Text to be shown below the panel containing the dotplots}

\item{rowSort}{if \code{TRUE} genes will be sorted according to the test statistic}

\item{colSort}{if \code{FALSE} chips will be sorted according to the phenotype}
}
\description{
Draws a heatmap of a matrix (e.g. gene expression data) together with
dotplots of a statistic (e.g. log fold change between experimental settings)
for each line
}
\details{
We call the dotplots \sQuote{sliders} because the faintly resemble sliders.
}
\examples{

    mat <- matrix(rnorm(120),ncol=6)
    stat <- rnorm(20)
    labels <- factor(rep(0,1,each=3))
    hcols <- gray.colors(32)
    lcols <- c('blue','red')

    heatslide(mat,stat,labels,hcols,lcols)

}
\author{
Florian Klinglmueller \email{float@lefant.net}
}

