% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/markdown.R
\name{html_statement}
\alias{html_statement}
\title{HTML Statistical Statement}
\usage{
html_statement(toc = FALSE)
}
\arguments{
\item{toc}{Whether to inclued a table of contents}
}
\description{
Custom format for statistical statements
}
\author{
float
}

