% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{list_fold}
\alias{list_fold}
\title{List fold (reduce)}
\usage{
list_fold(list, fun, z = NULL)
}
\arguments{
\item{list}{}

\item{fun}{a binary function taking elements of the list as values}

\item{z}{initial value, if empty the first element of the list will be used}
}
\value{
same as output of \code{fun}
}
\description{
Reduce for lists
}
\author{
float
}

