% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makeUCSClink}
\alias{makeUCSClink}
\title{UCSC Browser link}
\usage{
makeUCSClink(locus, project = "", link = "UCSC")
}
\arguments{
\item{locus}{transcript location in the form "loc:start-end"}

\item{project}{url string giving the project page at ucsc}
}
\value{
a character string with the url to the tag in the ucsc
}
\description{
Make a link to the projects UCSC Browser page
}
\author{
float
}

