% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{move_to_front}
\alias{move_to_front}
\title{Move columns to front}
\usage{
move_to_front(DT, cols)
}
\arguments{
\item{DT}{\code{data.table} object}

\item{cols}{names of columns to move to the front}
}
\value{
\code{data.table} with rearranged columns
}
\description{
Rearrange columns of a \code{data.table} object, moving a list of columns to the front, while leaving the remaining columns in their original order.
}
\author{
Florian Klinglmueller
}

