% Generated by roxygen2 (4.0.0): do not edit by hand
\name{myTopTags}
\alias{myTopTags}
\title{Top Tags from edgeR analysis}
\usage{
myTopTags(lrts, N = NULL, annodata = NULL, key = "nearest_ref_id")
}
\arguments{
\item{lrts}{DGELRT object}

\item{N}{number of tags to display (if not set will show all values with FDR adjusted p-value below 0.05)}

\item{annodata}{a data.table object with additionall annotation data}

\item{key}{index variable shared with annotation data}
}
\value{
a data.table with the top expressed tags
}
\description{
Makes a nice table from edger test results
}
\examples{
## setting up some random data
data <- matrix(rpois(600,200),ncol=6)
dge <- DGEList(counts=data)
dge <- estimateDisp(dge)

## a design matrix
design <- data.frame('Treatment' = factor(rep(0:1,each=3)))
mm <- model.matrix(~0+Treatment,design)

## a contrast matrix
colnames(mm) <- c('control','treatment')
cm <- makeContrasts('control-treatment',levels=mm)

## fit and test
out <- fit.edger(dge,mm,cm)

myTopTags(out,3)
}
\author{
Florian Klinglmueller
}

