% Generated by roxygen2 (4.0.0): do not edit by hand
\name{newest_vfile}
\alias{newest_vfile}
\title{Newest (versioned) file}
\usage{
newest_vfile(base, path = getwd(), ending = "Rd", format = "\%y\%m\%d",
  ostime = FALSE)
}
\arguments{
\item{base}{character giving the main part of the filename}

\item{path}{path to directory to search in}

\item{ending}{file ending}

\item{format}{format for date time version}

\item{ostime}{(defunct) should the time of file generation return by the OS
be used (useful for hashed files)?}
}
\description{
Searches a directory for versioned files with the given base name and
returns its newest filename.
}

