% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{normDist}
\alias{normDist}
\title{Distance between raw and normalized data}
\usage{
normDist(raw, normalized)
}
\arguments{
\item{raw}{\code{expressionSet} with non-normalized data}

\item{normalized}{\code{expressionSet} with normalized data}
}
\value{
numeric vector with euclidean distances
}
\description{
Compute euclidean distance between raw and normalized data for each probe.
}
\author{
float
}

