% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pdDist}
\alias{pdDist}
\title{paired distances}
\usage{
pdDist(raw, normalized)
}
\arguments{
\item{raw}{first numeric matrix}

\item{normalized}{second numeric matrix}
}
\value{
numeric vector with distances
}
\description{
Computed euclidean distance between corresponding rows of two matrices
}
\author{
float
}

