% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{plot_limma}
\alias{plot_limma}
\title{Plot Limma}
\usage{
plot_limma(model, coef, data, pheno, anno, top = 50,
  subset = 1:length(pheno))
}
\arguments{
\item{model}{\code{MArrayLM} object fit by \code{limma}}

\item{coef}{\code{integer} index of contrast to sort genes by}

\item{data}{\code{ExpressionSet} with normalized expression values}

\item{pheno}{\code{factor} with phenotype labels}

\item{anno}{\code{ProbeAnnDbBimap} of probe annotations}

\item{top}{\code{integer} number of genes to plot}

\item{subset}{\code{integer} vector of column numbers to be plotted}
}
\value{
plots a heatslide plot
}
\description{
Make a heatmap from the results of a limma fit
}
\author{
float
}

