% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotNormDist}
\alias{plotNormDist}
\title{Plot normDist}
\usage{
plotNormDist(raw, normalized, ...)
}
\arguments{
\item{raw}{\code{expressionSet} with raw data}

\item{normalized}{\code{expressionSet} with normalized data}

\item{...}{further arguments to \code{\link{normDist}}}
}
\value{
ggplot2 plot
}
\description{
Plot distribution of distances between raw and normalized data - potentially for a number of subsets
}
\author{
float
}

