% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger.R
\name{plotTopTags}
\alias{plotTopTags}
\title{TopTag heatmap}
\usage{
plotTopTags(tt, groups, dataCols, anno = "gene_short_name")
}
\arguments{
\item{tt}{data.table results from myTopTags}

\item{groups}{sample groups to be outlined in the plot}

\item{dataCols}{columns with data to be shown}

\item{anno}{columns with tag identifiers}
}
\value{
grid plot of heatmap
}
\description{
Make a heatmap from myTopTags
}
\author{
float
}

