% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.tbl_df}
\alias{print.tbl_df}
\title{Print tbl_df objects}
\usage{
\method{print}{tbl_df}(x, ..., round = 3, n = NULL, width = NULL)
}
\arguments{
\item{x}{Object to show}

\item{round}{Number of digits to print after the comma}

\item{n}{Number of rows to show. If ‘NULL’, the default, will print all rows if less than option ‘dplyr.print_max’. Otherwise, will print ‘dplyr.print_min’}

\item{width}{Width of text output to generate. This defaults to NULL, which means use ‘getOption("width")’ and only display the columns that fit on one screen. You can also set ‘options(dplyr.width = Inf)’ to override this default and always print all columns.}
}
\description{
Modified print function for tbl_df that permits rounding of numeric entries
}
\details{
Taken and modified from Richard Scrivens answere to http://stackoverflow.com/questions/34246552/
}
\author{
float
}

