% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{print_limma}
\alias{print_limma}
\title{print limma table}
\usage{
print_limma(model, coef, data, genenames, symbols, lookup = NULL, top = Inf)
}
\arguments{
\item{model}{Limma fitted model}

\item{coef}{vector of coefficients for which statst should be included}

\item{data}{\code{ExpressionSet} with data}

\item{genenames}{\code{ProbeAnnDbBimap} with long gene descriptions}

\item{symbols}{\code{ProbeAnnDbBimap} with short gene descriptions}

\item{lookup}{named vector lookup table that translates contrast defs into language}
}
\value{
\code{data.frame} with annotated statistics
}
\description{
Make a nice table for multiple limma estimated contrasts
}
\author{
float
}

