% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readRandom}
\alias{readRandom}
\title{read random numbers}
\usage{
readRandom(n, a, b, dev = "/dev/urandom")
}
\arguments{
\item{n}{number of random numbers to read}

\item{a}{start of range}

\item{b}{end of range}

\item{dev}{device file location}
}
\value{
numeric vector of leng \code{n} with numbers between \code{a} and \code{b}
}
\description{
Read random numbers from sytems random devices
}
\details{
Based on stackoverflow question http://stackoverflow.com/questions/11505039
}
\author{
Spacedman
}

