% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replicate2}
\alias{replicate2}
\title{parallel replicate with progress}
\usage{
replicate2(n, expr)
}
\arguments{
\item{n}{integer: the number of replications.}

\item{expr}{the expression (a language object, usually a call) to evaluate repeatedly.}
}
\value{
An atomic vector or matrix or list of the same length as \code{n}.
}
\description{
‘replicate’ is a wrapper for the common use of ‘mclapply2’ for
    repeated evaluation of an expression (which will usually involve
    random number generation).
}
\author{
float
}

