% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Rtop}
\alias{Rtop}
\title{Print system load and user overview}
\usage{
Rtop()
}
\description{
This script uses top to print the system load, memory usage and a list of
active users with their proportion of memory and cpu usage.
}

