% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sd_confint}
\alias{sd_confint}
\title{Standard deviation confidence interval}
\usage{
sd_confint(sd, n, alpha)
}
\arguments{
\item{sd}{estimated standard deviation}

\item{n}{number of observations used for the estimate}

\item{alpha}{desired coverage probability}
}
\value{
numeric vector with lower and upper bound
}
\description{
Computes a confidence interval for the standard deviation for a
standard deviation estimate
}
\author{
float
}

