% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{setwd}
\alias{setwd}
\title{modified setwd}
\usage{
setwd(dir)
}
\arguments{
\item{dir}{A character string}
}
\description{
Does windows to linux and vice versa home directory name
translation
}
\details{
If the detected OS is Linux directories starting with '//' will be renamed to '~/' if it is Windows the otherway around, else the directory name stays unchanged
}
\author{
float
}
\seealso{
\code{\link[base]{setwd}} for the base function
}

