% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{vfile}
\alias{vfile}
\title{Make a versioned filename}
\usage{
vfile(base, ending = "Rd", format = "\%y\%m\%d", hash = FALSE)
}
\arguments{
\item{base}{character giving the main part of the filename}

\item{ending}{file ending}

\item{format}{format for date time version}

\item{hash}{(defunct) if true a hash is used for versioning the filename}
}
\description{
This function returns a filename with a version extension to avoid
overwriting older results. The extension may either be the date (with time)
or a random hash. The latter may be useful if you want to obscure the date
of the data generation.
}

