% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoe.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(mat, highlight = 50, anno = 10)
}
\arguments{
\item{mat}{annotated data matrix, needs to have columns P.Value, logFC, and GeneSymbol}

\item{highlight}{How many dots in red}

\item{anno}{How many dots with annotations}
}
\value{
plots a volcano plot
}
\description{
Makes a volcano plot
}
\author{
Florian Klinglmueller
}

