% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{bt88.03.704-package}
\alias{bt88.03.704-package}
\title{float's utility functions

Some useful little functions that make life easier.}
\description{
float's utility functions

Some useful little functions that make life easier.
}

