% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{corner}
\alias{corner}
\title{Show upper left corner of a matrix}
\usage{
corner(dat, n = 6, m = 4)
}
\arguments{
\item{dat}{A matrix or dataframe}

\item{n}{Number of rows to be shown at most}

\item{m}{Number of columns to be shown at most}
}
\value{
A matrix or data.frame with \code{n} rows and \code{m} columns
}
\description{
This function prints the upper left corner of a matrix.
}
\examples{
corner(matrix(rnorm(1000),100,100))
}
\author{
float
}

