% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{groupwiseScatter}
\alias{groupwiseScatter}
\title{groupwise Scatter plots}
\usage{
groupwiseScatter(data, pheno, faulty = NULL)
}
\arguments{
\item{data}{\code{expressionSet} with data to plot}

\item{pheno}{treatment factor variable}

\item{faulty}{optional - columns to be highlighted (e.g. chips with quality issues)}
}
\value{
ggplot2 object
}
\description{
Draws MA (or Bland-Altman) plots for all pairs of arrays within groups of a treatment factor
}
\author{
float
}

