% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/heatslide.R
\name{heatslide}
\alias{heatslide}
\title{Heatmap with dotplot like visualization of effect sizes}
\usage{
heatslide(mat, stat, pheno, genenames = NULL,
  hcols = colorRampPalette(c("blue", "white", "red"))(32),
  lcols = rainbow(length(levels(pheno))), scale = c("row", "column",
  "none"), slidetitle = "Log (Base 2) Foldchange")
}
\arguments{
\item{mat}{Matrix containing the values to be plotted in the heatmap}

\item{stat}{Vector of values that should be plotted by the dotplot beside
each line of the heatmap}

\item{pheno}{Vector giving the group labels for each column of the matrix}

\item{genenames}{Character vector giving genenames (or corresponding annotation)}

\item{hcols}{Colours to be used in the heatmap}

\item{lcols}{Colours to indicate the column labels}

\item{scale}{Scale expression values by either 'row', 'column', or 'none'}

\item{slidetitle}{Text to be shown below the panel containing the dotplots}
}
\description{
Draws a heatmap of a matrix (e.g. gene expression data) together with
dotplots of a statistic (e.g. log fold change between experimental settings)
for each line
}
\details{
We call the dotplots \sQuote{sliders} because the faintly resemble sliders.
}
\examples{
mat <- matrix(rnorm(120),ncol=6)
    stat <- rnorm(20)
    labels <- factor(rep(0,1,each=3))
    hcols <- gray.colors(32)
    lcols <- c('blue','red')

    heatslide(mat,stat,labels,hcols,lcols)
}
\author{
Florian Klinglmueller \email{float@lefant.net}
}

