% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{html_statement}
\alias{html_statement}
\title{HTML Statistical Statement}
\usage{
html_statement(toc = FALSE, final = FALSE)
}
\arguments{
\item{toc}{Whether to inclued a table of contents}

\item{final}{Whether still in draft (will print more warnings etc.)}
}
\description{
Custom format for statistical statements
}
\author{
float
}

