% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/edger.R
\name{makeUCSClink}
\alias{makeUCSClink}
\title{UCSC Browser link}
\usage{
makeUCSClink(locus, project = "&hgsid=201952662_blwe2mLIAlJ066Rb2rEAvZDqZiGz")
}
\arguments{
\item{locus}{transcript location in the form "loc:start-end"}

\item{project}{url string giving the project page at ucsc}
}
\value{
a character string with the url to the tag in the ucsc
}
\description{
Make a link to the projects UCSC Browser page
}
\author{
float
}

