% Generated by roxygen2 (4.0.0): do not edit by hand
\name{normDist}
\alias{normDist}
\title{Distance between raw and normalized data}
\usage{
normDist(raw, normalized, rowsets = NULL, colsets = NULL)
}
\arguments{
\item{raw}{\code{expressionSet} with non-normalized data}

\item{normalized}{\code{expressionSet} with normalized data}

\item{rowsets}{a named \code{list} of row index sets for which the distances should be computed}

\item{colsets}{a named \code{list} of column index sets for which the distances should be computed}
}
\value{
numeric vector with euclidean distances
}
\description{
Compute euclidean distance between raw and normalized data for each probe.
}
\author{
float
}

