% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{pdDist}
\alias{pdDist}
\title{paired distances}
\usage{
pdDist(raw, normalized)
}
\arguments{
\item{raw}{first numeric matrix}

\item{normalized}{second numeric matrix}
}
\value{
numeric vector with distances
}
\description{
Computed euclidean distance between corresponding rows of two matrices
}
\author{
float
}

