% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoe.R
\name{volcanoe_plot}
\alias{volcanoe_plot}
\title{Volcanoe plot}
\usage{
volcanoe_plot(mat, highlight = 50, anno = 10)
}
\arguments{
\item{mat}{annotated data matrix, needs to have columns P.Value, logFC, and GeneSymbol}

\item{highlight}{How many dots in red}

\item{anno}{How many dots with annotations}
}
\value{
plots a volcanoe plot
}
\description{
Makes a volcanoe plot
}
\author{
Florian Klinglmueller
}

